Fixes issue with plugs firing more than once per action
Plugs were firing twice because we were applying the route combinator after we tested executed the plug. This means that if multiple actions used POST, and both actions had the same plug, the plug would until it found a handler that matched on the route.

In the case of Create we yield two HttpHandlers. The first one matches when the Path is "" and the second is when the path is "/". If we do a POST to "/" the plug would fire for the first handler, and then the 2nd handler. If we do a POST to "" then the first handler would be called and the plug would only fire once. This issue was fixed by normalizing the path when we want to support either an empty string or "/".