Fix for plugs firing twice (#143)
* WIP: Prevent plug from firing twice when Update action is set

What's happening?

1. AddHandler (or AddKeyHandler) is called for each action. The AddHandler method composes the specified plugs with each HTTP verb.
2. The Update action gets wired up with plugs for POSTs to resource/id and for PUTs
3. The Create action gets wired up with plugs for POSTs to resource

if you look at lines 234/235 we add plugs for create for POST. This is fine

We then do this again for at line 241 for Update when Update.IsSome (line 238). NOW we have the plug applied for POST twice.

Ideally we should only apply the plug to a method once. Talking to Krzystof last week, he was hesistant about changing the methods (or adding separate methods for applying the plugs). We could keep track of whether or not a plug has been applied for a method though, and not apply it if it's already been applied. Since the entire handler is composed with a verb at the same time the plugs are applied, we would need to associate the plugs with an action, and then associate the action with the handler, and only associate the plug with the action once. This may mean keeping track of separate handlers for the verb portion, applying the plugs, and then composing the action/handler with the underlying handler associated with that action.

* Test fails even when update isn't there

* Fixes issue with plugs firing more than once per action

Plugs were firing twice because we were applying the route combinator after we tested executed the plug. This means that if multiple actions used POST, and both actions had the same plug, the plug would until it found a handler that matched on the route.

In the case of Create we yield two HttpHandlers. The first one matches when the Path is "" and the second is when the path is "/". If we do a POST to "/" the plug would fire for the first handler, and then the 2nd handler. If we do a POST to "" then the first handler would be called and the plug would only fire once. This issue was fixed by normalizing the path when we want to support either an empty string or "/".