Allow all functions in a channel to see the socketId
Breaking change. Could be done in a none breaking way, by adding new operators to all code to opt-in to see the socketId, but adds considerable complexity.
Added a SocketInfo record to wrap the SocketId, to allow more information about the socket to be added at a later date without a breaking change.